share progress bar b/w regular and mega gam
